Moves ComplianceRequestTable to own component
